feat: 모임게시판/커뮤니티게시판 첫화면 컴포넌트화
