+++ client/src/components/Comment.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { useForm, SubmitHandler } from 'react-hook-form';
-import { CommentInput } from '../pages/CommunityDetail';
+import { CommentInput } from '../pages/ContentsDetail';
 interface CommentProps {
     commentData: {
         memberId: number;
