+++ client/src/pages/ContentsDetail.tsx
